refactor(workflow): eliminate worker-db package and simplify workflows
