--- packages/worker-db/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@shumai/worker-db",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
-  "dependencies": {
-    "@shumai/db": "workspace:*",
-    "@shumai/dtos": "workspace:*",
-    "@shumai/core": "workspace:*",
-    "@shumai/workflow-core": "workspace:*",
-    "@temporalio/activity": "^1.16.1",
-    "@temporalio/worker": "^1.16.1",
-    "@temporalio/workflow": "^1.16.1",
-    "ulid": "^3.0.2",
-    "jittered-fractional-indexing": "^1.0.1",
-    "@earendil-works/pi-agent-core": "0.78.0"
-  }
-}
