--- packages/worker-db/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "ignoreDeprecations": "6.0"
-  },
-  "include": ["src/**/*"]
-}
